+++ src/input_mocking.rs
@@ -49,8 +49,7 @@ use crate::user_input::*;
 ///
 /// let mut app = App::new();
 ///
-/// // This functionality requires Bevy's InputPlugin to be present in your plugin set.
-/// // If you included the `DefaultPlugins`, then InputPlugin is already included.
+/// // This functionality requires Bevy's InputPlugin (included with DefaultPlugins)
 /// app.add_plugins(InputPlugin);
 ///
 /// // Press a key press directly.
@@ -186,22 +185,69 @@ pub trait MockInput {
     fn reset_inputs(&mut self);
 }
 
-/// Query [`ButtonInput`] state directly for testing purposes.
+/// Query input state directly for testing purposes.
 ///
 /// In game code, you should (almost) always be using [`ActionState`](crate::action_state::ActionState)
 /// methods instead.
+///
+/// # Examples
+///
+/// ```rust
+/// use bevy::prelude::*;
+/// use bevy::input::InputPlugin;
+/// use leafwing_input_manager::input_mocking::QueryInput;
+/// use leafwing_input_manager::prelude::*;
+///
+/// let mut app = App::new();
+///
+/// // This functionality requires Bevy's InputPlugin (included with DefaultPlugins)
+/// app.add_plugins(InputPlugin);
+///
+/// // Check if a key is currently pressed down.
+/// let pressed = app.pressed(KeyCode::KeyB);
+///
+/// // Read the current vertical mouse scroll value.
+/// let value = app.read_axis_values(SingleAxis::mouse_wheel_y());
+///
+/// // Read the current changes in relative mouse X and Y coordinates.
+/// let values = app.read_axis_values(DualAxis::mouse_motion());
+/// let x = values[0];
+/// let y = values[1];
+/// ```
 pub trait QueryInput {
-    /// Checks if the `input` is currently pressed.
+    /// Checks if the `input` is currently pressed or active.
     ///
-    /// This method is intended as a convenience for testing; check the [`ButtonInput`] resource directly,
-    /// or use an [`InputMap`](crate::input_map::InputMap) in real code.
+    /// This method is intended as a convenience for testing;
+    /// use an [`InputMap`](crate::input_map::InputMap) in real code.
     fn pressed(&self, input: impl Into<UserInput>) -> bool;
 
-    /// Checks if the `input` is currently pressed the specified [`Gamepad`].
+    /// Checks if the `input` is currently pressed or active on the specified [`Gamepad`].
     ///
-    /// This method is intended as a convenience for testing; check the [`ButtonInput`] resource directly,
-    /// or use an [`InputMap`](crate::input_map::InputMap) in real code.
+    /// This method is intended as a convenience for testing;
+    /// use an [`InputMap`](crate::input_map::InputMap) in real code.
     fn pressed_on_gamepad(&self, input: impl Into<UserInput>, gamepad: Option<Gamepad>) -> bool;
+
+    /// Retrieves the values on all axes represented by the `input`.
+    ///
+    /// Binary inputs (e.g., keys and buttons) are treated like single-axis inputs,
+    /// typically returning a value between `0.0` (not pressed) and `1.0` (fully pressed).
+    ///
+    /// This method is intended as a convenience for testing;
+    /// use an [`InputMap`](crate::input_map::InputMap) in real code.
+    fn read_axis_values(&self, input: impl Into<UserInput>) -> Vec<f32>;
+
+    /// Retrieves the values on all axes represented by the `input` on the specified [`Gamepad`].
+    ///
+    /// Binary inputs (e.g., keys and buttons) are treated like single-axis inputs,
+    /// typically returning a value between `0.0` (not pressed) and `1.0` (fully pressed).
+    ///
+    /// This method is intended as a convenience for testing;
+    /// use an [`InputMap`](crate::input_map::InputMap) in real code.
+    fn read_axis_values_on_gamepad(
+        &self,
+        input: impl Into<UserInput>,
+        gamepad: Option<Gamepad>,
+    ) -> Vec<f32>;
 }
 
 /// Send fake UI interaction for testing purposes.
@@ -402,15 +448,39 @@ impl MutableInputStreams<'_> {
 }
 
 impl QueryInput for InputStreams<'_> {
+    #[inline]
     fn pressed(&self, input: impl Into<UserInput>) -> bool {
         self.input_pressed(&input.into())
     }
 
+    #[inline]
     fn pressed_on_gamepad(&self, input: impl Into<UserInput>, gamepad: Option<Gamepad>) -> bool {
         let mut input_streams = self.clone();
         input_streams.associated_gamepad = gamepad;
 
-        input_streams.input_pressed(&input.into())
+        input_streams.pressed(input)
+    }
+
+    #[inline]
+    fn read_axis_values(&self, input: impl Into<UserInput>) -> Vec<f32> {
+        let input = input.into();
+
+        if let Some(data) = self.input_axis_pair(&input) {
+            return vec![data.x(), data.y()];
+        }
+
+        vec![self.input_value(&input)]
+    }
+
+    fn read_axis_values_on_gamepad(
+        &self,
+        input: impl Into<UserInput>,
+        gamepad: Option<Gamepad>,
+    ) -> Vec<f32> {
+        let mut input_streams = self.clone();
+        input_streams.associated_gamepad = gamepad;
+
+        input_streams.read_axis_values(input)
     }
 }
 
@@ -513,7 +583,21 @@ impl QueryInput for World {
     fn pressed_on_gamepad(&self, input: impl Into<UserInput>, gamepad: Option<Gamepad>) -> bool {
         let input_streams = InputStreams::from_world(self, gamepad);
 
-        input_streams.input_pressed(&input.into())
+        input_streams.pressed(input)
+    }
+
+    fn read_axis_values(&self, input: impl Into<UserInput>) -> Vec<f32> {
+        self.read_axis_values_on_gamepad(input, None)
+    }
+
+    fn read_axis_values_on_gamepad(
+        &self,
+        input: impl Into<UserInput>,
+        gamepad: Option<Gamepad>,
+    ) -> Vec<f32> {
+        let input_streams = InputStreams::from_world(self, gamepad);
+
+        input_streams.read_axis_values(input)
     }
 }
 
@@ -584,6 +668,18 @@ impl QueryInput for App {
     fn pressed_on_gamepad(&self, input: impl Into<UserInput>, gamepad: Option<Gamepad>) -> bool {
         self.world.pressed_on_gamepad(input, gamepad)
     }
+
+    fn read_axis_values(&self, input: impl Into<UserInput>) -> Vec<f32> {
+        self.world.read_axis_values(input)
+    }
+
+    fn read_axis_values_on_gamepad(
+        &self,
+        input: impl Into<UserInput>,
+        gamepad: Option<Gamepad>,
+    ) -> Vec<f32> {
+        self.world.read_axis_values_on_gamepad(input, gamepad)
+    }
 }
 
 #[cfg(feature = "ui")]
@@ -600,18 +696,34 @@ impl MockUIInteraction for App {
 #[cfg(test)]
 mod test {
     use crate::input_mocking::{MockInput, QueryInput};
-    use bevy::{
-        input::{
-            gamepad::{GamepadConnection, GamepadConnectionEvent, GamepadEvent, GamepadInfo},
-            InputPlugin,
-        },
-        prelude::*,
+    use crate::prelude::{DualAxis, SingleAxis};
+    use bevy::input::gamepad::{
+        GamepadConnection, GamepadConnectionEvent, GamepadEvent, GamepadInfo,
     };
+    use bevy::input::InputPlugin;
+    use bevy::prelude::*;
+
+    fn test_app() -> App {
+        let mut app = App::new();
+        app.add_plugins(InputPlugin);
+
+        let gamepad = Gamepad::new(0);
+        let mut gamepad_events = app.world.resource_mut::<Events<GamepadEvent>>();
+        gamepad_events.send(GamepadEvent::Connection(GamepadConnectionEvent {
+            gamepad,
+            connection: GamepadConnection::Connected(GamepadInfo {
+                name: "TestController".into(),
+            }),
+        }));
+        app.update();
+        app.update();
+
+        app
+    }
 
     #[test]
     fn ordinary_button_inputs() {
-        let mut app = App::new();
-        app.add_plugins(InputPlugin);
+        let mut app = test_app();
 
         // Test that buttons are unpressed by default
         assert!(!app.pressed(KeyCode::Space));
@@ -640,47 +752,30 @@ mod test {
 
     #[test]
     fn explicit_gamepad_button_inputs() {
-        let mut app = App::new();
-        app.add_plugins(InputPlugin);
-
-        let gamepad = Gamepad { id: 0 };
-        let mut gamepad_events = app.world.resource_mut::<Events<GamepadEvent>>();
-        gamepad_events.send(GamepadEvent::Connection(GamepadConnectionEvent {
-            gamepad,
-            connection: GamepadConnection::Connected(GamepadInfo {
-                name: "TestController".into(),
-            }),
-        }));
-        app.update();
+        let mut app = test_app();
+        let gamepad = Some(Gamepad::new(0));
 
         // Test that buttons are unpressed by default
-        assert!(!app.pressed_on_gamepad(GamepadButtonType::North, Some(gamepad)));
+        assert!(!app.pressed_on_gamepad(GamepadButtonType::North, gamepad));
 
         // Press buttons
-        app.press_input_as_gamepad(GamepadButtonType::North, Some(gamepad));
+        app.press_input_as_gamepad(GamepadButtonType::North, gamepad);
         app.update();
+
+        // Verify the button are pressed
+        assert!(app.pressed_on_gamepad(GamepadButtonType::North, gamepad));
 
         // Test that resetting inputs works
         app.reset_inputs();
         app.update();
 
-        assert!(!app.pressed_on_gamepad(GamepadButtonType::North, Some(gamepad)));
+        // Verify the button are released
+        assert!(!app.pressed_on_gamepad(GamepadButtonType::North, gamepad));
     }
 
     #[test]
     fn implicit_gamepad_button_inputs() {
-        let mut app = App::new();
-        app.add_plugins(InputPlugin);
-
-        let gamepad = Gamepad { id: 0 };
-        let mut gamepad_events = app.world.resource_mut::<Events<GamepadEvent>>();
-        gamepad_events.send(GamepadEvent::Connection(GamepadConnectionEvent {
-            gamepad,
-            connection: GamepadConnection::Connected(GamepadInfo {
-                name: "TestController".into(),
-            }),
-        }));
-        app.update();
+        let mut app = test_app();
 
         // Test that buttons are unpressed by default
         assert!(!app.pressed(GamepadButtonType::North));
@@ -689,14 +784,54 @@ mod test {
         app.press_input(GamepadButtonType::North);
         app.update();
 
-        // Test the convenient `pressed` API
+        // Verify the button are pressed
         assert!(app.pressed(GamepadButtonType::North));
 
         // Test that resetting inputs works
         app.reset_inputs();
         app.update();
 
+        // Verify the button are released
         assert!(!app.pressed(GamepadButtonType::North));
+    }
+
+    #[test]
+    fn mouse_inputs() {
+        let mut app = test_app();
+
+        // Mouse axes should be inactive by default (no scroll or movement)
+        assert_eq!(app.read_axis_values(DualAxis::mouse_wheel()), [0.0, 0.0]);
+        assert_eq!(app.read_axis_values(DualAxis::mouse_motion()), [0.0, 0.0]);
+
+        // Send a simulated mouse scroll event with a value of 3 (positive for up)
+        app.send_axis_values(SingleAxis::mouse_wheel_y(), [3.0]);
+        app.update();
+
+        // Verify the mouse wheel Y axis reflects the simulated scroll
+        assert_eq!(app.read_axis_values(SingleAxis::mouse_wheel_y()), [3.0]);
+
+        // Verify other axes aren't affected
+        assert_eq!(app.read_axis_values(SingleAxis::mouse_wheel_x()), [0.0]);
+        assert_eq!(app.read_axis_values(DualAxis::mouse_motion()), [0.0, 0.0]);
+
+        // Send a simulated mouse movement event with a delta of (3.0, 2.0)
+        app.send_axis_values(DualAxis::mouse_motion(), [3.0, 2.0]);
+        app.update();
+
+        // Verify the mouse motion axes reflects the simulated movement
+        assert_eq!(app.read_axis_values(DualAxis::mouse_motion()), [3.0, 2.0]);
+
+        // Mouse input data is typically reset every frame
+        // Verify other axes aren't affected
+        assert_eq!(app.read_axis_values(DualAxis::mouse_wheel()), [0.0, 0.0]);
+
+        // Test that resetting inputs works
+        app.reset_inputs();
+        app.update();
+
+        // Verify all axes have no value after reset
+        assert_eq!(app.read_axis_values(SingleAxis::mouse_wheel_y()), [0.0]);
+        assert_eq!(app.read_axis_values(DualAxis::mouse_motion()), [0.0, 0.0]);
     }
 
     #[test]